% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Create a desktop notification}
\usage{
notify(subject, body = NULL)
}
\arguments{
\item{subject,body}{Message contents}
}
\description{
This function creates a desktop notification, currently using the
\code{notify-send} command.
}

